refactor: move issue activity tracking from django channels to rqworker
