+++ apiserver/Procfile
@@ -1,3 +1,2 @@
 web: gunicorn -w 4 -k uvicorn.workers.UvicornWorker plane.asgi:application --bind 0.0.0.0:$PORT --config gunicorn.config.py --max-requests 10000 --max-requests-jitter 1000 --access-logfile -
 worker: python manage.py rqworker
-channel-worker: python manage.py runworker issue-activites
