--- apiserver/bin/channel-worker
@@ -1,6 +1,0 @@
-#!/bin/bash
-set -e
-
-python manage.py wait_for_db
-python manage.py migrate
-python manage.py runworker issue-activites
