--- apiserver/plane/api/consumers/__init__.py
@@ -1,1 +1,0 @@
-from .issue_consumer import IssueConsumer
