+++ apiserver/plane/asgi.py
@@ -1,6 +1,6 @@
 import os
 
-from channels.routing import ProtocolTypeRouter, ChannelNameRouter
+from channels.routing import ProtocolTypeRouter
 from django.core.asgi import get_asgi_application
 
 django_asgi_app = get_asgi_application()
@@ -10,15 +10,9 @@ os.environ.setdefault("DJANGO_SETTINGS_MODULE", "plane.settings.production")
 # Initialize Django ASGI application early to ensure the AppRegistry
 # is populated before importing code that may import ORM models.
 
-from plane.api.consumers import IssueConsumer
 
 application = ProtocolTypeRouter(
     {
         "http": get_asgi_application(),
-        "channel": ChannelNameRouter(
-            {
-                "issue-activites": IssueConsumer.as_asgi(),
-            }
-        ),
     }
 )
