+++ apiserver/plane/settings/local.py
@@ -66,11 +66,3 @@ RQ_QUEUES = {
 
 WEB_URL = "http://localhost:3000"
 
-CHANNEL_LAYERS = {
-    "default": {
-        "BACKEND": "channels_redis.core.RedisChannelLayer",
-        "CONFIG": {
-            "hosts": [(REDIS_HOST, REDIS_PORT)],
-        },
-    },
-}
