+++ apiserver/plane/settings/staging.py
@@ -1,11 +1,8 @@
 """Production settings and globals."""
-import ssl
-from typing import Optional
 from urllib.parse import urlparse
 
 import dj_database_url
 from urllib.parse import urlparse
-from redis.asyncio.connection import Connection, RedisSSLContext
 
 import sentry_sdk
 from sentry_sdk.integrations.django import DjangoIntegration
@@ -184,53 +181,6 @@ RQ_QUEUES = {
     "default": {
         "USE_REDIS_CACHE": "default",
     }
-}
-
-class CustomSSLConnection(Connection):
-    def __init__(
-        self,
-        ssl_context: Optional[str] = None,
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.ssl_context = RedisSSLContext(ssl_context)
-
-class RedisSSLContext:
-    __slots__ = (
-        "context",
-    )
-
-    def __init__(
-        self,
-        ssl_context,
-    ):
-        self.context = ssl_context
-
-    def get(self):
-        return self.context
-
-
-url = urlparse(os.environ.get("REDIS_URL"))
-
-ssl_context = ssl.SSLContext()
-ssl_context.check_hostname = False
-
-CHANNEL_LAYERS = {
-    'default': {
-        'BACKEND': 'channels_redis.core.RedisChannelLayer',
-        'CONFIG': {
-            'hosts': [
-                    {
-                        'host': url.hostname,
-                        'port': url.port,
-                        'username': url.username,
-                        'password': url.password,
-                        'connection_class': CustomSSLConnection,
-                        'ssl_context': ssl_context,
-                    }
-                ],
-        }
-    },
 }
